backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
GeoAlchemy2==0.14.6
python-dotenv==1.0.1